Image Generator/requirements.txt
# Original dependencies
spotipy==2.23.0
python-dotenv==1.0.0
flask==3.0.0
torch==2.1.0
torchvision==0.16.0
diffusers==0.24.0
transformers==4.35.0
accelerate==0.24.0
pillow==10.1.0
numpy==1.24.3

